Replace double with float

+++ Imagine.Models/Color.cs
@@ -1,12 +1,12 @@
 namespace Imagine.Models;
 
-public readonly record struct Color(double R, double G, double B)
+public readonly record struct Color(float R, float G, float B)
 {
-	public static readonly Color Black = new(0D, 0D, 0D);
-	public static readonly Color Blue = new(0D, 0D, 1D);
-	public static readonly Color Green = new(0D, 1D, 0D);
-	public static readonly Color Red = new(1D, 0D, 0D);
-	public static readonly Color White = new(1D, 1D, 1D);
+	public static readonly Color Black = new(0F, 0F, 0F);
+	public static readonly Color Blue = new(0F, 0F, 1F);
+	public static readonly Color Green = new(0F, 1F, 0F);
+	public static readonly Color Red = new(1F, 0F, 0F);
+	public static readonly Color White = new(1F, 1F, 1F);
 
 	public static explicit operator Rgba32(Color value) =>
 		new(ToByte(value.R), ToByte(value.G), ToByte(value.B));
@@ -14,15 +14,15 @@ public readonly record struct Color(double R, double G, double B)
 	public static Color Average(List<Color> colors) =>
 		new(colors.Average(color => color.R), colors.Average(color => color.G), colors.Average(color => color.B));
 
-	public static Color operator *(Color left, double right) =>
+	public static Color operator *(Color left, float right) =>
 		new(left.R * right, left.G * right, left.B * right);
 
-	private static byte ToByte(double value)
+	private static byte ToByte(float value)
 	{
-		value = double.Clamp(value, 0D, 1D);
+		value = float.Clamp(value, 0F, 1F);
 
-		var result = double.Floor((1D + byte.MaxValue) * value);
-		result = double.Clamp(result, byte.MinValue, byte.MaxValue);
+		var result = float.Floor((1F + byte.MaxValue) * value);
+		result = float.Clamp(result, byte.MinValue, byte.MaxValue);
 
 		return (byte)result;
 	}
