+++ Imagine.Models/ProjectorSettings.cs
@@ -3,17 +3,17 @@ namespace Imagine.Models;
 public readonly record struct ProjectorSettings(
 	Vector3 Eye,
 	Vector3 Focus,
-	double HalfOpeningAngle,
+	float HalfOpeningAngle,
 	Color BackgroundColor)
 {
-	public static ProjectorSettings WithOpeningRadius(Vector3 eye, Vector3 focus, double openingRadius, Color backgroundColor) =>
+	public static ProjectorSettings WithOpeningRadius(Vector3 eye, Vector3 focus, float openingRadius, Color backgroundColor) =>
 		new(
 			Eye: eye,
 			Focus: focus,
-			HalfOpeningAngle: double.Atan2(openingRadius, (focus - eye).Length()),
+			HalfOpeningAngle: float.Atan2(openingRadius, (focus - eye).Length()),
 			BackgroundColor: backgroundColor);
 
-	public static ProjectorSettings WithOpeningRadius(Vector3 eye, Vector3 focus, double openingRadius) =>
+	public static ProjectorSettings WithOpeningRadius(Vector3 eye, Vector3 focus, float openingRadius) =>
 		WithOpeningRadius(
 			eye: eye,
 			focus: focus,
